+++ Data/SBV/SMT/SMTLib2.hs
@@ -125,7 +125,7 @@ cvt ctx kindInfo isSat comments (inputs, trackerVars) skolemInps (allConsts, con
                                 , "***    Found declaration: " ++ unwords (map getN defs)
                                 , "***    Quantified args  : " ++ unwords (map pretty foralls)
                                 , "***"
-                                , "*** If you use smtFunction/smtRecFunction/addAxiom, you cannot have explicit quantifiers."
+                                , "*** If you use smtFunction/addAxiom, you cannot have explicit quantifiers."
                                 , "*** Please report this as a feature request, if you cannot work around this limitation."
                                 ]